requirements.txt
streamlit
matplotlib
numpy
scipy
scikit-image
pyyaml
opencv-python  # if using OpenCV for saving video
